# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CategorySelectionRequest": {
        "description": "Parameters for selecting category sets for a room/game.",
        "properties": {
          "count": {
            "default": 1,
            "minimum": 1.0,
            "title": "Count",
            "type": "integer"
          },
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "locale": {
            "$ref": "#/components/schemas/LanguageCode",
            "default": "en"
          },
          "locales": {
            "items": {
              "$ref": "#/components/schemas/LanguageCode"
            },
            "title": "Locales",
            "type": "array"
          },
          "player_count": {
            "default": 2,
            "minimum": 1.0,
            "title": "Player Count",
            "type": "integer"
          }
        },
        "required": ["difficulty"],
        "title": "CategorySelectionRequest",
        "type": "object"
      },
      "CategorySelectionResponse": {
        "description": "Response for room-scoped category selection.",
        "properties": {
          "difficulty": {
            "$ref": "#/components/schemas/Difficulty"
          },
          "selections": {
            "items": {
              "$ref": "#/components/schemas/SelectedCategorySet"
            },
            "title": "Selections",
            "type": "array"
          }
        },
        "required": ["difficulty", "selections"],
        "title": "CategorySelectionResponse",
        "type": "object"
      },
...
      "Difficulty": {
        "enum": ["easy", "medium", "hard"],
        "type": "string"
      },
      "GamePhase": {
        "description": "Valid room lifecycle phases for game flow state.",
        "enum": ["lobby", "drawing", "guessing", "round_results", "final_results"],
        "title": "GamePhase",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "LanguageCode": {
        "pattern": "^[a-z]{2,5}(?:-[A-Za-z0-9]{2,8})?$",
        "type": "string"
      },
...
      "RoomStatusResponse": {
        "description": "Response for room status checks.",
        "properties": {
          "exists": {
            "title": "Exists",
            "type": "boolean"
          },
          "game_phase": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GamePhase"
              },
              {
                "type": "null"
              }
            ]
          },
          "players": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Players"
          }
        },
        "required": ["exists"],
        "title": "RoomStatusResponse",
        "type": "object"
      },
      "SelectedCategorySet": {
        "description": "One selected localized category set for game setup.",
        "properties": {
          "alternatives": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Alternatives",
            "type": "object"
          },
          "category_id": {
            "title": "Category Id",
            "type": "integer"
          },
          "category_name": {
            "title": "Category Name",
            "type": "string"
          },
          "item_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Item Ids",
            "type": "array"
          },
          "items": {
            "items": {
              "type": "string"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": ["category_id", "category_name", "item_ids", "items", "alternatives"],
        "title": "SelectedCategorySet",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Real-time multiplayer drawing game backend",
    "title": "Six Second Scribbles API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/rooms/{room_id}/category-selection": {
      "post": {
        "description": "Select category sets for a specific room/game setup.",
        "operationId": "select_room_categories_api_rooms__room_id__category_selection_post",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategorySelectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategorySelectionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Room Categories",
        "tags": ["rooms"]
      }
    },
    "/api/rooms/{room_id}/status": {
      "get": {
        "description": "Get the status of a room.",
        "operationId": "room_status_api_rooms__room_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "room_id",
            "required": true,
            "schema": {
              "title": "Room Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Room Status",
        "tags": ["rooms"]
      }
    },
...
  }
}